streamlit_ui/requirements.txt
streamlit == 1.27.0
streamlit-authenticator == 0.2.3
PyYAML == 6.0.1